week 3

+++ run.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
-rm -rf /tmp/processing
-mkdir /tmp/processing
-processing-java --output=/tmp/processing/ --force --sketch=$1 --run
+
+if [[ "$1" == "" ]]; then
+	echo The arguement is empty!
+else
+	rm -rf /tmp/processing
+	mkdir /tmp/processing
+	processing-java --sketch="$(pwd)/$1" --output=/tmp/processing --force --run
+fi
